% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fortel.R
\name{fortel}
\alias{fortel}
\title{Digital Fortune Teller}
\usage{
fortel()
}
\arguments{
\item{colour}{Pick a colour}
}
\value{
your forune
}
\description{
This function simulates the origami fortune teller game
}
\examples{
fortel()
}
\keyword{fortune}
