% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fortuneteller.R
\docType{package}
\name{fortuneteller}
\alias{fortuneteller}
\title{Digital Fortune Teller
Contains one function which simulates the fortune teller origami game}
\author{
Adeline Torey \email(adeline.torey@gmail.com)
}
